Added frontend for user pro's calendar selection

+++ src/Controller/EventController.php
@@ -3,10 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Event;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
 
@@ -16,10 +16,17 @@ use Symfony\Component\Serializer\SerializerInterface;
 class EventController extends AbstractController
 {
     /**
+     * Used in AJAX call in the calendar edit.
+     *
      * @Route("/{id}", name="event_edit", methods={"PUT"})
+     * @IsGranted("ROLE_PRO")
      */
     public function editEvent(Request $request, Event $event, SerializerInterface $serializer): JsonResponse
     {
+        if ($event->getCalendar()->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
+
         $json = json_decode($request->getContent(), false);
         $event->setStart(new \DateTimeImmutable($json->start));
         $event->setEnd(new \DateTimeImmutable($json->end));
@@ -31,10 +38,17 @@ class EventController extends AbstractController
     }
 
     /**
+     * Used in AJAX call in the calendar edit.
+     *
      * @Route("/{id}", name="event_delete", methods={"DELETE"})
+     * @IsGranted("ROLE_PRO")
      */
     public function deleteEvent(Event $event): JsonResponse
     {
+        if ($event->getCalendar()->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException();
+        }
+
         $entityManager = $this->getDoctrine()->getManager();
         $entityManager->remove($event);
         $entityManager->flush();
